+++ apps/backend/src/notifications/notifications.module.ts
@@ -1,0 +1,25 @@
+import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { BullModule } from '@nestjs/bullmq';
+import { NotificationsService } from './notifications.service';
+import { NotificationsController } from './notifications.controller';
+import { NotificationsProcessor } from './notifications.processor';
+import { Notification } from './entities/notification.entity';
+import { NotificationTemplate } from './entities/notification-template.entity';
+import { MailerModule } from '../mailer/mailer.module';
+import { UsersModule } from '../users/users.module';
+
+@Module({
+  imports: [
+    TypeOrmModule.forFeature([Notification, NotificationTemplate]),
+    BullModule.registerQueue({
+      name: 'notifications',
+    }),
+    MailerModule,
+    UsersModule,
+  ],
+  controllers: [NotificationsController],
+  providers: [NotificationsService, NotificationsProcessor],
+  exports: [NotificationsService],
+})
+export class NotificationsModule {}